requirements.txt
django
django-autocomplete-light
pillow
django-phonenumber-field
django-widget-tweaks==1.4.3
django-extra-views==0.11.0
